fix: config validator checks for conflict with bounding and graphtran… (#352)
## Description
This PR introduces a new validator in the config to check whether the
config.model.decoder.initialise_data_extractor_zero = True is not used
with any bounding.

Closes #346 

## What problem does this change solve?
This is mainly a concern for the GraphTransformer in the decoder, see
issue #346.

## What issue or task does this change relate to?
#346

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

---------

Co-authored-by: Jakob Schloer <jakob.schloer@gmail.com>