feat(graphs): build nodes from xarray compatible data (#330)
## Description

This PR adds an XArray node builder for building nodes from files
compatible with xarray.

Example usage: 
```
    data_nodes = XArrayNodes("/storage/danra/height_levels.zarr", "data", lat_key='latitude', lon_key='longitude')
    graph = HeteroData()
    graph = data_nodes.update_graph(graph)
```
Builds a set of nodes on the latitudes and longitudes specified in the
file. Therefore this builders requires latitudes and longitudes to be
defined on the file. The names of these can be specified with the
`lat/lon_key` kwargs, defaulting to `lat` and `lon`, such that xarray
can access them with its `__getitem__` method on the instantiated
`<xarrray.Dataset>`.

Additionally, since this work is related to zarr, some mock objects
(that were anemoi-dataset rather than zarr mocks) called for example
`mock_zarr_dataset`, have been renamed to `mock_anemoi_dataset`, such
that the `mock_zarr_dataset` is made available for actual zarr dataset
mocks in the test suite.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Mario Santa Cruz <48736305+JPXKQX@users.noreply.github.com>