test: add test to restart training from existing checkpoint (#381)
## Description
This PR adds a test to restart training from an existing checkpoint.

## What problem does this change solve?
It is not aimed to be a blocking test but a first step to help track
changes that result in incompatibilities with existing checkpoints which
could lead to problems with e.g. transfer learning.

## What issue or task does this change relate to?
Related to the discussion about transfer learning but not directly
addressing the issue discussed.