fix(graphs): revert PR #379 (#409)
## Description
<!-- What issue or task does this change relate to? -->

Revert PR #379 .

Thanks to @havardhhaugen for spotting this.

## What problem does this change solve?
<!-- Describe if it's a bugfix, new feature, doc update, or breaking
change -->

Raising an error during graph creation will break multi-GPU training.
Currently, one limitation of `anemoi-graphs` is that it does not support
distributed creation. For example, during multi-GPU training with
on-the-fly graph creation, the same graph is created on all GPUs. In
this case, raising an error will cause the code to fail on some of the
GPUs.

## What issue or task does this change relate to?
<!-- link to Issue Number -->

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>