fix(graphs): undo masking when torch-cluster is installed (#375)
## Description
<!-- What issue or task does this change relate to? -->
The current workflow did not function as intended with `torch-cluster`
when a source or target mask was specified. The `torch-cluster` library
is used only when graph creation needs to be optimized. It is not listed
as a dependency because it is difficult to install in the CI
environment, and as a result, it is only lightly tested.

## What problem does this change solve?
<!-- Describe if it's a bugfix, new feature, doc update, or breaking
change -->
This PR fixes the previously reported bug and introduces a new base
class shared by both KNNEdges and CutOffEdges to handle masking.

## What issue or task does this change relate to?
<!-- link to Issue Number -->

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***


<!-- readthedocs-preview anemoi-training start -->
----
📚 Documentation preview 📚:
https://anemoi-training--375.org.readthedocs.build/en/375/

<!-- readthedocs-preview anemoi-training end -->

<!-- readthedocs-preview anemoi-graphs start -->
----
📚 Documentation preview 📚:
https://anemoi-graphs--375.org.readthedocs.build/en/375/

<!-- readthedocs-preview anemoi-graphs end -->

<!-- readthedocs-preview anemoi-models start -->
----
📚 Documentation preview 📚:
https://anemoi-models--375.org.readthedocs.build/en/375/

<!-- readthedocs-preview anemoi-models end -->

---------

Co-authored-by: Dieter Van den Bleeken <dieter.vandenbleeken@meteo.be>