fix: mlflow hp params limit (#424)
## Description
When logging to many hp_params into the mlflow server, the runs and the
experiment takes long to load. We set a configurable limit to better
control this, so if a run has many hp_params the code will crash and
avoid that run being logged. This limit can be configured from the
config, but if not limit is passed then the default value of 2000 will
be used.

Additionally this PR modifies the W&B schema so the identity when W&B is
not enabled, is update to None so it doesn't crash or complaint about
the '?'

## What problem does this change solve?
<!-- Describe if it's a bugfix, new feature, doc update, or breaking
change -->

## What issue or task does this change relate to?
<!-- link to Issue Number -->

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)