feat(models): postprocessor for nans in diagnostic fields (#461)
## Description
Place NaNs in the output of diagnostic variables at locations of NaNs of
another variable.

## What problem does this change solve?
The new logic of the imputer takes the NaN locations from the input and
places them in the output. This is not possible for diagnostic variables
as they are not present in the input.

## What issue or task does this change relate to?
Part of https://github.com/ecmwf/anemoi-core/issues/370


***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)


<!-- readthedocs-preview anemoi-training start -->
----
📚 Documentation preview 📚:
https://anemoi-training--461.org.readthedocs.build/en/461/

<!-- readthedocs-preview anemoi-training end -->

<!-- readthedocs-preview anemoi-graphs start -->
----
📚 Documentation preview 📚:
https://anemoi-graphs--461.org.readthedocs.build/en/461/

<!-- readthedocs-preview anemoi-graphs end -->

<!-- readthedocs-preview anemoi-models start -->
----
📚 Documentation preview 📚:
https://anemoi-models--461.org.readthedocs.build/en/461/

<!-- readthedocs-preview anemoi-models end -->

---------

Co-authored-by: Harrison Cook <harrison.cook@ecmwf.int>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Vera Gahlen <vera.gahlen@ecmwf.int>