fix(training): callbacks when using rollout  (#490)
## Description

Since the imputer depends on the batchsize now, we need to first apply
the postprocessors, and then concatenate several forecast steps when
rollout is >1. This PR solves this.
See artifacts on mlflow here:
https://mlflow.ecmwf.int/#/experiments/335/runs/d568028a6ef941779f4fb7e3e0941f49/artifacts

At the moment, on main, the callbacks when using rollout > 1 and the
imputer fail with a size-missmatch error.

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>