feat: checkpoint migrations (#386)
New features or refactoring of models might break model checkpoints.
As discussed during a @ecmwf/anemoitechnicalsubgroup meeting, this PR
introduces a systematic way of migrating model checkpoints so that they
can be always up-to-date with the anemoi version. This will avoid the
hard choice of updating anemoi and retrain a checkpoint or keep an old
anemoi version.

## The idea
The idea is to defer the responsability of keeping a working checkpoint
to the person who introduces a breaking change.
This person only has to handle **his own** breaking change. This means
that he can expect that checkpoint are working before.

This takes inspiration from database migration systems that apply
changes to a database incrementally.

Specifically:
-  Checkpoints will have a new "migrations" key that store migrations
that were already executed (if the key does not exist, it is assumed to
not have any migration executed);
- when saving a checkpoint after training, the "migrations" key is
populated with all existing migrations (no need to execute them);
- if a PR introduces a change that breaks checkpoint, a new migration
must be created with:
```bash
anemoi-models migration create my-migration-name
```
Here is an example migration file:
```python
from anemoi.models.migrations import CkptType
from anemoi.models.migrations import MigrationMetadata

metadata = MigrationMetadata(
    versions={
        "migration": "1.0.0",  # this tracks the version of the migration script. Might be useful if we want to refactor the API
        "anemoi-models": "0.8.1",  # this tracks the version of anemoi-models when this migration was created
    }
)


def migrate(ckpt: CkptType) -> CkptType:
    """Migrate the checkpoint"""
    ckpt["test"] = ckpt.pop("baz")
    return ckpt


def rollback(ckpt: CkptType) -> CkptType:
    """Rollback the migration"""
    ckpt["baz"] = ckpt.pop("test")
    return ckpt
```
This file contains some version metadata, a `migrate` function that
should fix the introduced breaking change of the checkpoint, and a
`rollback` function that does the opposite opperation.

## User interface
The user can use the command
```bash
anemoi-models migration sync ckpt_path [--steps STEPS]
```
to execute migrations of an old checkpoint. Only missing migrations will
be executed.

Additional option:
- `--steps STEPS`: will execute only STEPS migration files. If negative,
will rollback STEPS times

## The pipeline
- A new PR indroduces a change that breaks checkpoints
- A CI pipeline detects that checkpoints are broken by this PR
- A new migration file must be created and filled
- CI passes again
- The PR can be merged
- Users will migrate their checkpoint with `anemoi-models migrate`
 
## Roadmap
Here is the progress status of this PR (we may want to split this for
the CI check?):
- [x] a script to generate new migrations
- [x] a system to select the migrations or rollbacks to execute
- [x] callback to store existing migrations when saving checkpoints
- [x] CI to detect that a PR broke a checkpoint and ask the developper
to create a new migration. This can be a complement to #381.
- [x] add docs (see tests for now for examples)


## Open Questions
- Should the migrator be automatically ran if a checkpoint is out of
date? Or should it always be done manually by the users? I tend to think
that it should be manual, with anemoi logging the command to run to
update the checkpoint when crashing.
- How to handle inference checkpoints? (probably regenerate the
inference checkpoint from the complete model after migration)
- How to handle different model types? Should there be a set of
migrations per models? Or a unique set of migrations but migrations also
take some additional parameters to know the model to migrate? (although
it can probably be infered from the checkpoint)

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***


<!-- readthedocs-preview anemoi-training start -->
----
📚 Documentation preview 📚:
https://anemoi-training--386.org.readthedocs.build/en/386/

<!-- readthedocs-preview anemoi-training end -->

<!-- readthedocs-preview anemoi-graphs start -->
----
📚 Documentation preview 📚:
https://anemoi-graphs--386.org.readthedocs.build/en/386/

<!-- readthedocs-preview anemoi-graphs end -->

<!-- readthedocs-preview anemoi-models start -->
----
📚 Documentation preview 📚:
https://anemoi-models--386.org.readthedocs.build/en/386/

<!-- readthedocs-preview anemoi-models end -->

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>