ci: Add checkpoint migrations workflow (#492)
## Description
The upcoming checkpoint migrations rely on the migration files having a
reference to the next anemoi-models version that the migration is
targeting. This is done with a placeholder
`%NEXT_ANEMOI_MODELS_VERSION%` that is replaced on release time with the
version being released.

This workflow does that. It runs on the release-please (DeployDuck) PRs,
and it updates the placeholder in all the migration files and commits it
to the release-please branch. If there are no migrations to update, it
does nothing. Since release-please regularly force-pushes its branch to
the latest state of main, any new migrations made since the opening of
the PR will be automatically updated.

When the release PR is merged, the migrations go into main with the
correct versions.

I made it a general workflow where we can put all future things to do at
release time.

## What issue or task does this change relate to?
#386


***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)