fix: one-variable model (#478)
## Description

The current version of anemoi fails when models with just one prognostic
variable are trained because of some .squeeze statements that reduce
arrays to scalars. I use .reshape(-1) instead. This seems to fix the
issue in my tests and still works with more complex models with multiple
variables.

## What problem does this change solve?

Enables models with one prognostic variable to be trained in
anemoi-train. This is arguably an edge case, which should be enabled
nevertheless.

## What issue or task does this change relate to?

This change addresses issue #477. 

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Ana Prieto Nemesio <91897203+anaprietonem@users.noreply.github.com>
Co-authored-by: anaprietonem <ana.prietonemesio@ecmwf.int>