feat: target indices (#426)
## Description
This PR introduces support for a new category of input features called
targets, which are used to compute the loss but are neither predicted by
the model nor part of the diagnostic or prognostic features. It refines
how input and output features are handled during training, especially in
the context of loss computation and data indexing.

## What problem does this change solve?
Currently, the feature classification in the Anemoi training pipeline
assumes:

- Prognostic = features predicted by the model (i.e., diagnostic -
forcing)
- Forcing = input-only features
- Diagnostic = output-only features 

However, we now have use cases where certain features:
- Are used as input
- Are required to compute the loss
- Are not predicted by the model

To handle this, we introduce a new category called targets, defined as
features used to compute the loss but not part of the forecast state.

This update allows:
- Better separation of model inputs and loss-relevant features.
- More flexible handling of loss computation by explicitly reindexing
predicted and target variables per loss.
- Improved interpretability of feature roles in the pipeline.

## What issue or task does this change relate to?
N/A (internal feature enhancement)

## Additional notes
Prognostic features are now defined as: prognostic = diagnostic -
forcing - targets

A feature can simultaneously be a target and a forcing. Loss functions
can be updated to explicitly reindex predictions and targets based on
their respective configurations.
This improves modularity of the loss pipeline and prepares for future
loss extensions.

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)


<!-- readthedocs-preview anemoi-training start -->
----
📚 Documentation preview 📚:
https://anemoi-training--426.org.readthedocs.build/en/426/

<!-- readthedocs-preview anemoi-training end -->

<!-- readthedocs-preview anemoi-graphs start -->
----
📚 Documentation preview 📚:
https://anemoi-graphs--426.org.readthedocs.build/en/426/

<!-- readthedocs-preview anemoi-graphs end -->

<!-- readthedocs-preview anemoi-models start -->
----
📚 Documentation preview 📚:
https://anemoi-models--426.org.readthedocs.build/en/426/

<!-- readthedocs-preview anemoi-models end -->

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Ana Prieto Nemesio <91897203+anaprietonem@users.noreply.github.com>